Rename controllers to routers
Just aligning with the FastAPI naming convention.
